Move some utility functions to their own files
